Issue 1941: ValueHolders are consuming too much memory
Reduce the amount of ValueHolders needed per reference, by reimplementing QueryBasedValueHolder's methods and fields in UnitOfWorkQueryValueHolder instead of wrapping an instance of QueryBasedValueHolder.
Note that this will require changes to change detection as now the original object and its clone share the same ValueHolder.
For this UnitOfWorkQueryValueHolder will save the original value read from the DB in the new field originalValue.
Change detection would then have to compare originalValue to value.
Updated change detection is not part of this commit.

Signed-off-by: Patrick Schmitt <Patrick.Schmitt@cpb-software.com>